services/paddleocr_service/requirements.txt
fastapi
uvicorn
paddleocr
paddlepaddle
pillow
python-multipart